small tweaks, rename some things to ts
