+++ src/libs/quotes.ts
