+++ src/libs/static.ts
