+++ src/profile/LanguageSelector.js
@@ -3,7 +3,7 @@ import FormGroup from "react-bootstrap/lib/FormGroup";
 import ControlLabel from "react-bootstrap/lib/ControlLabel";
 import FormControl from "react-bootstrap/lib/FormControl";
 import { MdAutorenew } from "react-icons/md";
-import LanguageContext from "../LanguageContext";
+import LanguageContext from "../state/LanguageContext";
 import { availableLanguages, updateLanguage } from "../libs/languages";
 
 const LanguageSelector = (props) => {
